[refactoring] Create nand handler class
